fixed errors in layouts regarding package restructure
