+++ DonationTracker/app/src/main/java/com/example/kory/donationtracker/Controller/Home.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
-import com.example.kory.donationtracker.Models.UserFacade;
+import com.example.kory.donationtracker.Models.UserClasses.UserFacade;
 
 import com.example.kory.donationtracker.R;
