+++ DonationTracker/app/src/main/java/com/example/kory/donationtracker/Controller/Login.java
@@ -6,7 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
-import com.example.kory.donationtracker.Models.UserFacade;
+import com.example.kory.donationtracker.Models.UserClasses.UserFacade;
 
 import com.example.kory.donationtracker.R;
