+++ DonationTracker/app/src/main/java/com/example/kory/donationtracker/Models/LocationClasses/LocationFacade.java
@@ -1,0 +1,5 @@
+package com.example.kory.donationtracker.Models.LocationClasses;
+
+public class LocationFacade {
+
+}
